[CALCITE-7398] Incorrect int cast in VariantNonNull#cast for BIGINT

+++ core/src/main/java/org/apache/calcite/runtime/variant/VariantNonNull.java
@@ -286,7 +286,7 @@ public class VariantNonNull extends VariantSqlValue {
           break;
         }
         case BIGINT: {
-          long l = (int) value;
+          long l = (long) value;
           switch (type.getTypeName()) {
           case TINYINT:
           case SMALLINT:
